Created Hunter interface. Replaced old Hunter class by HunterPlayer. Adapted the rest of the software.
